requirements.txt
ultralytics>=8.0.0
opencv-python>=4.5.0
numpy>=1.20.0
PyYAML>=6.0
requests>=2.25.0